requirements.txt
torch
torchvision
Pillow
numpy
matplotlib
tqdm
arithmetic-compressor
bitarray